refactor: rename SeleniumConfiguration to Configuration
