--- eyes_common/applitools/common/config/__init__.py
@@ -1,11 +1,0 @@
-from .configuration import BatchInfo, Configuration
-from .misc import BrowserType, StitchMode
-from .selenium_configuration import SeleniumConfiguration
-
-__all__ = (
-    "Configuration",
-    "BatchInfo",
-    "StitchMode",
-    "BrowserType",
-    "SeleniumConfiguration",
-)
