+++ eyes_common/applitools/common/selenium/__init__.py
@@ -1,0 +1,4 @@
+from .config import Configuration
+from .misc import BrowserType, StitchMode
+
+__all__ = ("StitchMode", "BrowserType", "Configuration")
